+++ src/pages/Search/components/CurrentRegionButton/index.tsx
@@ -1,4 +1,4 @@
-import CustomButton from '@/components/CustomButton';
+import CustomButton from '@/components/CustomMui/CustomButton';
 import LocationIcon from '@/components/icon/Location';
 import { S } from './style';
 import useAppSelector from '@/hooks/useAppSelector';
